# docs/swagger.yaml
definitions:
...
  handlers.createStudentRequest:
    properties:
      course_id:
        type: string
      created_at:
        type: string
      crm_link:
        type: string
      curator_id:
        type: string
      full_name:
        type: string
      is_active:
        enum:
        - активен
        - неактивен
        example: активен
        type: string
      parent_name:
        type: string
      parent_phone_number:
        type: string
      phone_number:
        type: string
      platform_link:
        type: string
    type: object
  handlers.updateStudentRequest:
    properties:
      course_id:
        type: string
      created_at:
        type: string
      crm_link:
        type: string
      curator_id:
        type: string
      full_name:
        type: string
      is_active:
        enum:
        - активен
        - неактивен
        example: активен
        type: string
      parent_name:
        type: string
      parent_phone_number:
        type: string
      phone_number:
        type: string
      platform_link:
        type: string
    type: object
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
  models.Student:
    properties:
      course_id:
        type: string
      created_at:
        type: string
      crm_link:
        type: string
      curator_id:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: string
      parent_name:
        type: string
      parent_phone_number:
        type: string
      phone_number:
        type: string
      platform_link:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /settings/students:
    get:
      description: Возвращает список студентов с возможностью фильтрации
      parameters:
      - description: Поиск по ФИО
        in: query
        name: search
        type: string
      - description: Фильтр по ID курса
        format: uuid
        in: query
        name: course
        type: string
      - description: Фильтр по активности
        in: body
        name: is_active
        schema:
          enum:
          - активен
          - неактивен
          type: string
      - description: Фильтр по ID куратора
        format: uuid
        in: query
        name: curator_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список студентов
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить список студентов
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: |-
        Создает запись о студенте. Допустимые значения:
        - is_active: активен, неактивен
        - created_at: дата в формате DD.MM.YYYY
        - phone_number: международный формат (+7XXX...)
      parameters:
      - description: Данные студента
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.createStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ID созданного студента
          schema:
            properties:
              id:
                type: string
            type: object
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Создать нового студента
      tags:
      - Students
  /settings/students/{studentId}:
    delete:
      description: Удаляет запись о студенте из системы
      parameters:
      - description: UUID студента
        format: uuid
        in: path
        name: studentId
        required: true
        type: string
      responses:
        "204":
          description: Студент успешно удален
        "400":
          description: Неверный формат UUID
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Студент не найден
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Удалить студента
      tags:
      - Students
    get:
      description: Возвращает полную информацию о студенте по его ID
      parameters:
      - description: UUID студента
        format: uuid
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные студента
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Неверный формат UUID
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Студент не найден
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить данные студента
      tags:
      - Students
    put:
      consumes:
      - application/json
      description: |-
        Обновляет информацию о существующем студенте. Допустимые значения:
        - is_active: активен, неактивен
        - created_at: дата в формате DD.MM.YYYY
        - phone_number: международный формат (+7XXX...)
      parameters:
      - description: UUID студента
        format: uuid
        in: path
        name: studentId
        required: true
        type: string
      - description: Статус студента
        in: body
        name: is_active
        required: true
        schema:
          enum:
          - активен
          - неактивен
          type: string
      - description: Обновленные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.updateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Данные успешно обновлены
        "400":
          description: Неверный формат данных
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Студент не найден
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Обновить данные студента
      tags:
      - Students
...
swagger: "2.0"